Inheritance for geomechanics solvers are changed
